Extract memory mapper into separate class
